Adjust bottom navigation
