+++ screens/SettingsScreen.js
@@ -1,0 +1,20 @@
+import * as React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+
+export default function SettingsScreen() {
+  return (
+    <View style={styles.container} contentContainerStyle={styles.contentContainer}>
+        
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+  },
+  contentContainer: {
+    paddingTop: 15,
+  }
+});
